split to fuse.py and schedule.py [pr]
